requirements.txt
aiocontextvars==0.2.2
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
celery==4.4.7
certifi==2022.6.15
charset-normalizer==2.0.12
clinto==0.4.0
contextvars==2.4
Django==3.2.13
django-autoslug==1.9.8
django-celery-results==2.3.1
django-storages==1.12.3
idna==3.3
immutables==0.18
importlib-metadata==4.8.3
jsonfield==2.1.1
kombu==4.6.11
loguru==0.5.3
lxml==4.9.0
mobi==0.3.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tencentcloud-sdk-python-common==3.0.658
tencentcloud-sdk-python-tts==3.0.658
typing-extensions==4.1.1
urllib3==1.26.9
vine==1.3.0
wooey==0.12.0
zipp==3.6.0
